+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Supported copying images to clipboard. (#515)
 - Improved image gallery mode styling and user experience. (#520)
 - Supported customizing typo style. (#488)
+- Supported marking below/above as readed. (#390, #524)
 
 ### Changed